Fixed circle issue

+++ software/src/machine.h
@@ -113,7 +113,7 @@ public:
   uint8_t executeMovementCommand();
   void clearCurrentCmd() {  memset(&this->newCmd, 0, sizeof(MachineCommand_t) );  }
   uint8_t setMotorsSpeed(uint16_t newSpeed);
-  uint8_t performAxisLinearMovement_G00(Motor& inputMotor, int32_t newAxisPosition);
+  uint8_t performAxisLinearMovement_G00(Motor& inputMotor, float newAxisPosition);
   uint8_t performLinearInterpolation_G01(Point3d_float_t& p1);
   uint8_t performLinearInterpolation_G01_Optimized(Point3d_int32_t& p1);
   uint8_t performCircularArcInterpolation(Motor& axis_1, Motor& axis_2, uint8_t DIR);
